requirements.txt
# FFmpeg needs to be installed separately for OpenAI whisper
# Rust for openAI whisper
# python version 3.10

torch==2.6.0
setuptools-rust==1.10.2
openai-whisper==20240930
sentence-transformers==3.4.1
requests==2.32.3
Python-IO==0.3
